+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Autoscaler Change Log
 - Skip completed/failed pods when getting cluster state
 - Fix GCP machines specification with allocatable values instead of capacity
 - Add n1_highmem machines' specifications
+- Allow to set optimizer probing time
 
 0.7.0 November 4, 2020
 ------------------------